requirements.txt
PySide6
awsiotsdk
pytest